refactor: Organise translations in different files
Before, there were chaos under the config/locale dir, because
translations for the same part of a view or model were on different
files, not being consistent at all.

In order to achieve a better organisation whith the translation files,
several files have been created for replicating the structure of the
app.