fix: Fix password update if ID has expired
Because of the validations that the model Member has, when updating a
password of a user with the ID expired from
ProfileController#update_password, the process failed.

In order to solve this, a new method has been created, bypassing the ID
expiration date validation in order to save the password.

Fixes: #70