refactor: Make less database queriesin role allocs
This commit reduces the effective number of calls to the database when
showing all roles allocations. Before, some calls for have to be made in
order to show each member. Now, via a join query, it is only needed one
query to the database to get all the needed data.